+++ src/main/java/blackjack/domain/BlackJackManager.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Deck;
+import blackjack.domain.card.ParticipantCards;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
