+++ src/main/java/blackjack/domain/Participant.java
@@ -1,5 +1,8 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.ParticipantCards;
+
 import java.util.List;
 
 public abstract class Participant {
