+++ src/main/java/blackjack/domain/Player.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.card.ParticipantCards;
+
 public class Player extends Participant {
     protected Player(final ParticipantCards cards, final String name) {
         super(cards, name);
