+++ src/test/java/blackjack/domain/CardNumberTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.card.CardNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
