+++ src/test/java/blackjack/fixture/ParticipantCardsFixture.java
@@ -1,7 +1,7 @@
 package blackjack.fixture;
 
-import blackjack.domain.Card;
-import blackjack.domain.ParticipantCards;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.ParticipantCards;
 
 import java.util.List;
