+++ src/test/java/blackjack/fixture/ParticipantFixture.java
@@ -1,8 +1,8 @@
 package blackjack.fixture;
 
-import blackjack.domain.Card;
+import blackjack.domain.card.Card;
 import blackjack.domain.Participant;
-import blackjack.domain.ParticipantCards;
+import blackjack.domain.card.ParticipantCards;
 
 import java.util.List;
